Add user role management and enhance database connection logic
- Implemented `register_employee` function in `auth.controller.js` for employee registration.
- Modified `driver.controller.js` to include tenant-specific driver management.
- Enhanced `vehicle.controller.js` for tenant-specific vehicle management.
- Updated `location.controller.js` to use tenant-specific state and district models.
- Improved `connection.js` to include data seeding logic.
- Refactored `tenantMenager.service.js` to manage tenant-specific data.
- Updated middleware to handle user roles and permissions more effectively.
- Adjusted routes to accommodate new employee registration and tenant-specific operations.

+++ src/controllers/driver.controller.js
@@ -1,10 +1,12 @@
 import { Op } from "sequelize";
 import { asyncHandler } from "../utils/asyncHandler.js";
+import { rootDB } from "../db/tenantMenager.service.js";
+const { models } = await rootDB();
+const { Driver } = models;
 
 
 // GET request
-const driverList = asyncHandler(async (req, res) => {
-    const { Driver } = req.dbModels;
+const allDriverList = asyncHandler(async (req, res) => {
     try {
         const { id = "", license_no = "" } = req.query;
 
@@ -36,14 +38,34 @@ const driverList = asyncHandler(async (req, res) => {
     }
 });
 
+const myDriverList = asyncHandler(async (req, res) => {
+    try {
+        const dbName = req.headers["x-tenant-id"];
+        if(!dbName) return res.status(400).json({ succss: false, code: 400, message: "'x-tenant-id' header required!!!" });
+
+        const driverList = await Driver.findAll({
+            where: { owned_by: dbName }
+        });
+        if (!driverList) return res.status(500).json({ success: false, code: 500, message: "Not Found!!!" });
+
+        return res.status(200).json({ success: true, code: 200, message: "Fetched Successfully", data: driverList });
+
+    } catch (error) {
+        console.log(error);
+        return res.status(500).json({ success: false, code: 500, message: error.message });
+    }
+});
+
 const deleteDriver = asyncHandler(async (req, res) => {
-    const { Driver } = req.dbModels;
     try {
         const { id } = req.params;
-
         if (!id) return res.status(400).json({ success: false, code: 400, message: "Driver id required!!!" });
 
-        await Driver.destroy({ where: { id } });
+        const dbName = req.headers["x-tenant-id"];
+        if(!dbName) return res.status(400).json({ succss: false, code: 400, message: "'x-tenant-id' header required!!!" });
+
+        const isDeleted = await Driver.destroy({ where: { id, owned_by: dbName } });
+        if(!isDeleted) return res.status(403).json({ success: false, code: 403, message: "Not Possible!!!" });
 
         return res.status(200).json({ success: true, code: 200, message: "Record Deleted." });
 
@@ -55,11 +77,12 @@ const deleteDriver = asyncHandler(async (req, res) => {
 
 // POST request
 const addDriver = asyncHandler(async (req, res) => {
-    const { Driver } = req.dbModels;
     try {
         const { name = "", license_no = "", contact_no = "", address = "" } = req.body;
-
         if (!license_no) return res.status(400).json({ succss: false, code: 400, message: "License number is required!!!" });
+
+        const dbName = req.headers["x-tenant-id"];
+        if(!dbName) return res.status(400).json({ succss: false, code: 400, message: "'x-tenant-id' header required!!!" });
 
         const isDriverExists = await Driver.findOne({ where: { license_no } });
 
@@ -69,7 +92,8 @@ const addDriver = asyncHandler(async (req, res) => {
             name,
             license_no,
             contact_no,
-            address
+            address,
+            owned_by: dbName
         });
 
         return res.status(200).json({ success: true, code: 200, message: "Driver Added." });
@@ -81,14 +105,14 @@ const addDriver = asyncHandler(async (req, res) => {
 });
 
 const editDriver = asyncHandler(async (req, res) => {
-    const { Driver } = req.dbModels;
     try {
         const { name = "", license_no = "", contact_no = "", address = "" } = req.body;
-
         if (!license_no) return res.status(400).json({ succss: false, code: 400, message: "License number is required!!!" });
 
-        const isDriverExists = await Driver.findOne({ where: { license_no } });
+        const dbName = req.headers["x-tenant-id"];
+        if(!dbName) return res.status(400).json({ succss: false, code: 400, message: "'x-tenant-id' header required!!!" });
 
+        const isDriverExists = await Driver.findOne({ where: { license_no } });
         if (!isDriverExists) return res.status(409).json({ success: false, code: 409, message: `Driver with license number ${license_no} is not exists!!!` });
 
         let replace_object = {}
@@ -96,10 +120,11 @@ const editDriver = asyncHandler(async (req, res) => {
         if (contact_no) replace_object.contact_no = contact_no.trim();
         if (address) replace_object.address = address.trim();
 
-        await Driver.update(
+        const isDeleted = await Driver.update(
             replace_object,
-            { where: { license_no } }
+            { where: { license_no, owned_by: dbName } }
         );
+        if(!isDeleted) return res.status(403).json({ success: false, code: 403, message: "Not Possible!!!" });
 
         return res.status(200).json({ success: true, code: 200, message: "Details updated." });
 
@@ -110,4 +135,4 @@ const editDriver = asyncHandler(async (req, res) => {
 });
 
 
-export { addDriver, deleteDriver, editDriver, driverList }
+export { addDriver, deleteDriver, editDriver, allDriverList, myDriverList };
